refactor of handler structure
