+++ services/gateway/handler/account/model.go
@@ -1,0 +1,1 @@
+package account
