+++ services/gateway/handler/auth/model.go
@@ -1,0 +1,1 @@
+package auth
