--- services/gateway/handler/handler.go
@@ -1,602 +1,0 @@
-package handler
-
-import (
-	"net/http"
-
-	"github.com/gin-gonic/gin"
-	"github.com/google/uuid"
-	cerror "github.com/idlab-discover/kebeng/common/cerror"
-	accClient "github.com/idlab-discover/kebeng/services/account/client"
-	assertionClient "github.com/idlab-discover/kebeng/services/assertion/client"
-	"github.com/idlab-discover/kebeng/services/gateway/internal/auth"
-	"github.com/idlab-discover/kebeng/services/gateway/internal/config"
-	"github.com/idlab-discover/kebeng/services/gateway/internal/message"
-	"github.com/idlab-discover/kebeng/services/gateway/internal/middleware"
-	storeClient "github.com/idlab-discover/kebeng/services/store/client"
-	storepb "github.com/idlab-discover/kebeng/services/store/proto"
-	"github.com/sirupsen/logrus"
-	"gopkg.in/macaroon.v2"
-)
-
-// this file handles all the http requests and maps it to the correct client
-
-type Handler struct {
-	config          *config.Config
-	AccountClient   accClient.AccountClientInterface
-	StoreClient     storeClient.StoreClientInterface
-	AssertionClient *assertionClient.AssertionClient // TODO: change to interface
-}
-
-func NewHandler(accountClient accClient.AccountClientInterface, storeClient storeClient.StoreClientInterface, config *config.Config) *Handler {
-	return &Handler{
-		config:        config,
-		AccountClient: accountClient,
-		StoreClient:   storeClient,
-	}
-}
-
-// TODO: add middelware function that checks authentication with discharge macaroon (can't get discharge macaroon as of now, so don't know the format/what it contains)
-func (h *Handler) SetupEndpoints(r *gin.Engine) {
-	// NOTE: doesn't really check macaroons yet just sets email to test value
-	authGroup := r.Group("/dev/api")
-	authGroup.Use(middleware.AuthMiddleware())
-
-	r.POST("api/v1/snaps/auth/nonces", h.RequestStoreDeviceNonce)      // TODO
-	r.POST("api/v1/snaps/auth/sessions", h.RequestStoreDeviceSessions) // TODO
-	r.POST("/v2/snaps/refresh", h.RefreshSnap)                         // TODO
-	r.POST("/createAccount", h.createAccount)
-	r.POST("/dev/api/register-name/", h.RegisterSnapName)
-	r.POST("/dev/api/acl/", h.generateMacaroon)
-
-	authGroup.GET("/account/", h.getAccount)
-	authGroup.PATCH("/account", h.patchAccount)
-	r.POST("/dev/api/register-name-dispute/", h.RegisterSnapNameDispute)
-	r.POST("/dev/api/snaps/:snap_id/builds", h.ProcessSnapBuildAssertion)
-	r.POST("/dev/api/acl/verify/", h.verifyMacaroon) //TODO: implement correctly to many unknows of what has to be included and what not, need good source
-
-	r.GET("/v2/snaps/find", h.FindSnaps) // TODO
-}
-
-// TODO: Implement this function properly
-// Right now it's just a placeholder to make sure Snapcraft can be installed in the lxc container
-func (h *Handler) RequestStoreDeviceNonce(c *gin.Context) {
-	c.JSON(http.StatusOK, message.RequestStoreDeviceNonceRes{Nonce: "this-nonce"})
-}
-
-// TODO: Implement this function properly
-// Right now it's just a placeholder to make sure Snapcraft can be installed in the lxc container
-func (h *Handler) RequestStoreDeviceSessions(c *gin.Context) {
-	c.JSON(http.StatusOK, gin.H{"macaroon": "shakamacaroon"})
-}
-
-// TODO: Implement this function properly
-// Add all the cases for the different refresh actions
-func (h *Handler) RefreshSnap(c *gin.Context) {
-	el := cerror.NewErrorList()
-	var req message.RefreshSnapRequest
-	if err := c.ShouldBindJSON(&req); err != nil {
-		el.Add(cerror.BadRequest, cerror.FormatBindError(err))
-		c.JSON(el.GetHTTPStatus(), gin.H{"error_list": el})
-		return
-	}
-
-	var resp message.RefreshSnapResponses
-	for _, action := range req.Actions {
-		if action.Action == "install" {
-			logrus.Infof("%v", action)
-			res := h.refreshSnapInstall(c, action, el)
-			resp.Responses = append(resp.Responses, res)
-		} else {
-			el.Add(cerror.NotImplemented, "Action not implemented")
-			c.JSON(el.GetHTTPStatus(), gin.H{"error_list": el})
-			return
-		}
-	}
-
-	c.JSON(http.StatusOK, resp)
-}
-
-// TODO: Implement this function properly
-func (h *Handler) refreshSnapInstall(c *gin.Context, action *message.Action, el *cerror.ErrorList) *message.RefreshSnapResult {
-	var res message.RefreshSnapResult
-
-	// Get snap entry -> will return 1 snap entry
-	snapEntries := h.StoreClient.GetEntries(&storepb.GetEntriesRequest{
-		Entries: []*storepb.GetEntryRequest{
-			{
-				Name:                action.Name,
-				Id:                  action.SnapID,
-				PreloadAssociations: []string{"REVISIONS"},
-			},
-		},
-	})
-	if len(snapEntries.Errors) > 0 {
-		el.ExtendStoreError(snapEntries.Errors)
-		res.Result = nil
-		return &res
-	}
-
-	// Get the snap entry -> should only be 1
-	snapEntry := snapEntries.Entries[0]
-
-	// Get a set of all the architectures of the snapEntry's revisions
-	architectureSet := make(map[string]struct{})
-	for _, snapRevision := range snapEntry.Revisions {
-		for _, arch := range snapRevision.Architectures {
-			architectureSet[arch] = struct{}{}
-		}
-	}
-
-	// Convert the set to a slice
-	architectures := make([]string, 0, len(architectureSet))
-	for arch := range architectureSet {
-		architectures = append(architectures, arch)
-	}
-
-	latest_revision := float64(snapEntry.Revisions[len(snapEntry.Revisions)-1].Sequence)
-
-	// Get the publisher of the snap entry
-	publisher := h.AccountClient.GetAccountByID(snapEntry.PublisherId)
-
-	result := "install"
-
-	res.Result = &result
-	res.InstanceKey = &action.InstanceKey
-	res.SnapId = &snapEntry.Id
-	res.Name = &snapEntry.SnapName
-	res.Snap = &message.RefreshSnap{
-		Architectures: &architectures,
-		SnapId:        &snapEntry.Id,
-		Name:          &snapEntry.SnapName,
-		Publisher: &message.Publisher{
-			Username: publisher.Username,
-			ID:       publisher.Id,
-		},
-		Download: &message.Download{
-			URL:      nil, // TODO: implement
-			Sha3_384: nil, // TODO: implement
-			Size:     nil, // TODO: implement
-		},
-		Version:     nil, // TODO: implement
-		Revision:    &latest_revision,
-		Confinement: snapEntry.Confinement,
-		Type:        snapEntry.Type,
-		Base:        snapEntry.Base,
-	}
-	return &res
-}
-
-func (h *Handler) FindSnaps(c *gin.Context) {
-	el := cerror.NewErrorList()
-	var req message.FindSnapsRequest
-	if err := c.ShouldBindJSON(&req); err != nil {
-		el.Add(cerror.BadRequest, cerror.FormatBindError(err))
-		c.JSON(el.GetHTTPStatus(), gin.H{"error_list": el})
-		return
-	}
-}
-
-func (h *Handler) createAccount(c *gin.Context) {
-	var req message.CreateAccountRequest
-	el := cerror.NewErrorList()
-	if err := c.ShouldBindJSON(&req); err != nil {
-		el.Add(cerror.BadRequest, cerror.FormatBindError(err))
-		c.JSON(el.GetHTTPStatus(), gin.H{"error_list": el})
-		return
-	}
-	account := h.AccountClient.CreateAccount(req.DisplayName, req.Username, req.Email)
-	if len(account.Errors) > 0 {
-		el.ExtendAccountError(account.Errors)
-		c.JSON(el.GetHTTPStatus(), gin.H{"error_list": el})
-		return
-	}
-
-	c.JSON(http.StatusOK, message.CreateAccountResponse{Id: account.Id})
-}
-
-func (h *Handler) RegisterSnapName(c *gin.Context) {
-	el := cerror.NewErrorList()
-	var req message.RegisterSnapNameReq
-	if err := c.ShouldBindJSON(&req); err != nil {
-		el.Add(cerror.BadRequest, cerror.FormatBindError(err))
-		c.JSON(el.GetHTTPStatus(), gin.H{"error_list": el})
-		return
-	}
-
-	dryRun := c.Query("dry_run") == "1"
-	resp := h.StoreClient.RegisterSnapName(req.SnapName, req.IsPrivate, req.Store, dryRun)
-	if len(resp.Errors) > 0 {
-		el.ExtendStoreError(resp.Errors)
-		c.JSON(el.GetHTTPStatus(), gin.H{"error_list": el})
-		return
-	}
-
-	statusCode := http.StatusCreated
-	if dryRun {
-		statusCode = http.StatusOK
-	}
-
-	c.JSON(statusCode, message.RegisterSnapNameRes{SnapId: resp.Id, SnapName: req.SnapName})
-}
-
-// TODO: Implement this function
-// Reason for not being implemented yet:
-// There is currently no way to keep track of disputes (in the database),
-// disputes should be handled by a natural person, not by the system,
-// which is not possible at the moment.
-func (h *Handler) RegisterSnapNameDispute(c *gin.Context) {
-	el := cerror.NewErrorList()
-	el.Add(cerror.NotImplemented, "Not implemented")
-	c.JSON(el.GetHTTPStatus(), gin.H{"error_list": el})
-}
-
-func (h *Handler) generateMacaroon(c *gin.Context) {
-	el := cerror.NewErrorList()
-	var req *message.GenerateMacaroonRequest
-	if err := c.ShouldBindJSON(&req); err != nil {
-		el.Add(cerror.BadRequest, cerror.FormatBindError(err))
-		c.JSON(el.GetHTTPStatus(), gin.H{"error_list": el})
-		return
-	}
-
-	// validate input
-	auth.ValidateGenerateMacaroonRequest(req, el)
-	if len(*el) > 0 {
-		c.JSON(http.StatusBadRequest,
-			gin.H{
-				"error_list": el,
-			})
-		return
-	}
-
-	// check whether snapEntry exists else 404 not found
-	entries := make([]*storepb.GetEntryRequest, len(req.Packages))
-	for i, p := range req.Packages {
-		entries[i] = &storepb.GetEntryRequest{
-			Name: &p.Name,
-			Id:   &p.SnapId,
-		}
-	}
-
-	entriesResponse := h.StoreClient.GetEntries(&storepb.GetEntriesRequest{Entries: entries})
-	if len(entriesResponse.Errors) > 0 {
-		el.ExtendStoreError(entriesResponse.Errors)
-		c.JSON(el.GetHTTPStatus(), gin.H{"error_list": el})
-		return
-	}
-	// mimic set to prevent duplicate ids in macaroon
-	snapIDs := make(map[string]bool, len(entriesResponse.Entries))
-	for _, e := range entriesResponse.Entries {
-		snapIDs[e.Id] = true
-	}
-
-	macaroon := auth.GenerateMacaroon(c, req, snapIDs, h.config.MacaroonConfig)
-	if len(macaroon.Errors) > 0 {
-		el.Extend(macaroon.Errors)
-		c.JSON(el.GetHTTPStatus(), gin.H{"error_list": el})
-		return
-	}
-	c.JSON(http.StatusOK, gin.H{"macaroon": macaroon.Macaroon})
-}
-
-func (h *Handler) ProcessSnapBuildAssertion(c *gin.Context) {
-	el := cerror.NewErrorList()
-	var req *message.SnapBuildAssertionReq
-	if err := c.ShouldBindJSON(&req); err != nil {
-		el.Add(cerror.BadRequest, cerror.FormatBindError(err))
-		c.JSON(el.GetHTTPStatus(), gin.H{"error_list": el})
-		return
-	}
-	// SnapID is a path parameter
-	snapID := c.Param("snap_id")
-	if snapID == "" {
-		el.Add(cerror.BadRequest, "snap_id is required")
-		c.JSON(el.GetHTTPStatus(), gin.H{"error_list": el})
-		return
-	}
-
-	// Process the snap build assertion with the snapID and req
-	resp := h.AssertionClient.ProcessSnapBuildAssertion(req.Assertion)
-	if len(resp.Errors) > 0 {
-		el.ExtendAssertionError(resp.Errors)
-		//c.JSON(el.GetHTTPStatus(), gin.H{"error_list": el}) // This would be the prefered way to return the error, but the documentation handles this error differently
-		c.JSON(http.StatusBadRequest, gin.H{"succes": false, "cerror": el})
-		return
-	}
-
-	c.JSON(http.StatusOK, gin.H{
-		"headers": gin.H{
-			"authority-id":      resp.AuthorityId,
-			"grade":             resp.Grade,
-			"sign-key-sha3-384": resp.SignKeySha3_384,
-			"snap-id":           resp.SnapId,
-			"snap-sha3-384":     resp.SnapSha3_384,
-			"snap-size":         resp.SnapSize,
-			"timestamp":         resp.Timestamp,
-			"type":              resp.Type,
-		},
-	})
-}
-
-func (h *Handler) getAccount(c *gin.Context) {
-	el := cerror.NewErrorList()
-	email, isThere := c.Get("email")
-	if !isThere {
-		el.Add(cerror.BadRequest, "missing email")
-		c.JSON(el.GetHTTPStatus(), gin.H{"error_list": el})
-		return
-	}
-
-	account := h.AccountClient.GetAccountByEmail(email.(string))
-	if len(account.Errors) > 0 {
-		el.ExtendAccountError(account.Errors)
-		c.JSON(http.StatusInternalServerError, gin.H{"error_list": el})
-		return
-	}
-	accId, err := uuid.Parse(account.Id)
-	if err != nil {
-		// should never happen really
-		el.Add(cerror.InternalServerError, err.Error())
-		c.JSON(el.GetHTTPStatus(), gin.H{"error_list": el})
-		return
-	}
-
-	keys := h.AccountClient.GetAccountKeysByAccountID(account.Id)
-	if len(keys.Errors) > 0 {
-		el.ExtendAccountError(keys.Errors)
-		c.JSON(http.StatusInternalServerError, gin.H{"error_list": el})
-		return
-	}
-	// convert to message.AccountKey
-	messageKeys := make([]message.AccountKey, len(keys.Keys))
-	for i, k := range keys.Keys {
-		messageKeys[i] = message.AccountKey{
-			Name:            k.Name,
-			PublicKeySHA384: k.Sha3384,
-			Since:           k.Since.AsTime(),
-			Until:           k.Until.AsTime(),
-		}
-	}
-
-	// Get snaps
-
-	entries := h.StoreClient.GetEntriesByAccountID(account.Id)
-	if len(entries.Errors) > 0 {
-		el.ExtendStoreError(entries.Errors)
-		c.JSON(http.StatusInternalServerError, gin.H{"error_list": el})
-		return
-	}
-
-	// Get all revisions for every snapEntry
-	snapEntries := make([]*storepb.GetRevisionsByEntryIdRequest, len(entries.Entries))
-	for i, e := range entries.Entries {
-		snapEntries[i] = &storepb.GetRevisionsByEntryIdRequest{EntryId: e.Id}
-	}
-
-	// do 1 request where we get all the snapRevisions for every snapEntry
-	revisions := h.StoreClient.GetRevisionsByEntryIds(
-		&storepb.GetRevisionsByEntryIdRequests{
-			Requests: snapEntries,
-		})
-	if len(revisions.Errors) > 0 {
-		el.ExtendStoreError(revisions.Errors)
-		c.JSON(http.StatusInternalServerError, gin.H{"error_list": el})
-		return
-	}
-
-	// according to the docs a Publisher is only linked to a snapEntry
-	// get all publishers per snap
-	accountIds := make([]string, len(entries.Entries))
-	for i, e := range entries.Entries {
-		accountIds[i] = e.PublisherId
-	}
-	publishers := h.AccountClient.GetAccountsByIds(accountIds)
-	if len(publishers.Errors) > 0 {
-		el.ExtendAccountError(publishers.Errors)
-		c.JSON(http.StatusInternalServerError, gin.H{"error_list": el})
-		return
-	}
-
-	// Now we have all the data we need to fill in the snaps object
-	// start filling in
-	snaps := make(map[string]map[string]message.Snap)
-
-	// map publishers by ID for quick lookup
-	publisherMap := make(map[string]*message.Publisher)
-	for _, p := range publishers.Accounts {
-		publisherMap[p.Id] = &message.Publisher{
-			ID:          p.Id,
-			DisplayName: p.DisplayName,
-			Username:    p.Username,
-			Validation:  getString(p.Validation),
-		}
-	}
-
-	// map revisions by entry ID for quick lookup
-	revisionMap := make(map[string][]message.SnapRevision)
-	for _, revs := range revisions.Responses {
-		entryID := revs.EntryId
-		for _, rev := range revs.Revisions {
-			revisionMap[entryID] = append(revisionMap[entryID], message.SnapRevision{
-				Revision:      int(rev.Sequence),
-				Version:       rev.Version,
-				Status:        rev.Status,
-				Architectures: rev.Architectures,
-				// Channels:      rev.Channels, // TODO: see how to get them should not be in revision object
-			})
-		}
-	}
-
-	for _, e := range entries.Entries {
-		series := getString(e.Base) // Example: "16" //TODO: check if this is series normally but i think its the same as base?
-		snapName := e.SnapName      // Example: "hello-published"
-
-		// Ensure series exists in the map
-		if snaps[series] == nil {
-			snaps[series] = make(map[string]message.Snap)
-		}
-
-		// Get publisher
-		publisher, exists := publisherMap[e.PublisherId]
-		if !exists {
-			// should always exist normally otherwise something went wrong in previous processes
-			publisher = &message.Publisher{
-				ID:          e.PublisherId, // Fallback to ID only
-				DisplayName: "Unknown Publisher",
-				Username:    "unknown",
-				Validation:  "unverified",
-			}
-		}
-
-		// Get revisions for this snap entry
-		latestRevisions := revisionMap[e.Id]
-
-		// TODO: fix this
-		snap := message.Snap{
-			Status: getString(e.Status),
-			Price:  getFloat64(e.Price),
-			Since:  e.Since.AsTime(),
-			SnapID: e.Id,
-			// Store:           e.Store, // not yet implemented
-			Private:         getBool(e.Private), // this isn't the best yet but don't know what to do if its a nil value set to true or false?
-			IconURL:         e.IconUrl,
-			Publisher:       *publisher,
-			LatestComments:  []message.SnapComment{}, // No comments available yet
-			LatestRevisions: latestRevisions,
-		}
-
-		// Assign snap to the correct series and snap name
-		snaps[series][snapName] = snap
-	}
-
-	// not yet implemented don't support brandstores yet
-	// stores := h.AccountClient.GetStores(account.Id)
-
-	// leaving out the deprecated fields for now
-	resp := message.AccountResponse{
-		ID:          accId,
-		DisplayName: account.DisplayName,
-		Email:       account.Email,
-		Username:    account.Username,
-
-		AccountKeys: messageKeys,
-		Snaps:       snaps,
-		Stores:      nil,
-	}
-	c.JSON(http.StatusOK, resp)
-}
-
-func (h *Handler) patchAccount(c *gin.Context) {
-	el := cerror.NewErrorList()
-	email, ok := c.Get("email")
-	if !ok {
-		el.Add(cerror.BadRequest, "missing email")
-		c.JSON(el.GetHTTPStatus(), gin.H{"error_list": el})
-		return
-	}
-
-	rootMacaroon, ok := c.Get("macaroon")
-	if !ok {
-		el.Add(cerror.BadRequest, "missing macaroon")
-		c.JSON(el.GetHTTPStatus(), gin.H{"error_list": el})
-		return
-	}
-
-	// check permissions of macaroon to include "edit_account" permission
-	if !auth.HasPermission(rootMacaroon.(*macaroon.Macaroon), "edit_account") {
-		el.Add(cerror.ResourceForbidden, "missing permission to edit account")
-		c.JSON(el.GetHTTPStatus(), gin.H{"error_list": el})
-		return
-	}
-
-	var req *message.AccountPatchRequest
-	if err := c.ShouldBindJSON(&req); err != nil {
-		el.Add(cerror.BadRequest, cerror.FormatBindError(err))
-		c.JSON(el.GetHTTPStatus(), gin.H{"error_list": el})
-		return
-	}
-
-	// patch account by email
-	account := h.AccountClient.PatchAccountByEmail(email.(string), req.ShortNameSpace)
-	if len(account.Errors) > 0 {
-		el.ExtendAccountError(account.Errors)
-		c.JSON(el.GetHTTPStatus(), gin.H{"error_list": el})
-		return
-	}
-
-	c.JSON(http.StatusOK, gin.H{"success": true})
-}
-
-// TODO: implement correctly to many unknows of what has to be included and what not, need good source
-func (h *Handler) verifyMacaroon(c *gin.Context) {
-	// not implemented for now
-	c.JSON(http.StatusNotImplemented, gin.H{"error_list": cerror.NewError(cerror.NotImplemented, "not implemented too many unknowns of implementation")})
-	return
-	/*
-			var req *message.VerifyRequest
-			el := cerror.NewErrorList()
-			if err := c.ShouldBindJSON(&req); err != nil {
-				el.Add(cerror.BadRequest, cerror.FormatBindError(err))
-				c.JSON(el.GetHTTPStatus(),gin.H{"error_list": el,})
-				return
-			}
-
-		    // TODO: Don't know if you can get email from this actually or not (from discharge macaroon normally?)
-			userEmail := auth.VerifyAndGetEmail(h.config, el, req.AuthData.Authorization)
-		    // need userEmail to continue
-			if userEmail == nil {
-		        c.JSON(el.GetHTTPStatus(), gin.H{"error_list": el})
-		        return
-			}
-
-		    // TODO: change account client to use errorList
-			user, _ := h.AccountClient.GetAccountByEmail(*userEmail)
-		    if user == nil {
-		        el.Add(cerror.ResourceNotFound, fmt.Sprintf("user with email %s not found", *userEmail))
-		        c.JSON(el.GetHTTPStatus(), gin.H{"error_list": el})
-		        return
-		    }
-
-		    // TODO: check wtf these values have to be
-		    // probably get them from the macaroon
-		    _ = message.VerifyResponse{
-		        Allowed:               true, // idk what dertemines this
-		        DeviceRefreshRequired: false, // obsolote
-		        RefreshRequired:       false, // check with expiry time of macaroon
-		        Account: &message.VerifyAccount{
-		            Email:       user.Email,
-		            DisplayName: user.DisplayName,
-		            OpenId:      "oid1234", // what id is this?
-		            Verified:    true, // idk what this is
-		        },
-		        Device:      nil, // obsolete
-		        LastAuth:    "2016-05-26T12:53:23Z", // should be from macaroon
-		        Permissions: &[]string{"package_access", "package_manage", "package_push", "package_register", "package_release", "package_update"}, // should be from macaroon
-		        SnapIds:     nil, // should be from macaroon
-		        Channels:    nil, // should be from macaroon
-		    }
-		    return
-	*/
-}
-
-// Helper functions for safe pointer dereferencing.
-func getString(s *string) string {
-	if s == nil {
-		return ""
-	}
-	return *s
-}
-
-func getFloat64(f *float64) float64 {
-	if f == nil {
-		return 0.0
-	}
-	return *f
-}
-
-func getBool(b *bool) bool {
-	if b == nil {
-		return false
-	}
-	return *b
-}
