+++ services/gateway/handler/snap/model.go
@@ -1,0 +1,1 @@
+package snap
